+++ compiler-plugin/src/main/java/io/ballerina/stdlib/websocket/plugin/PluginConstants.java
@@ -34,6 +34,7 @@ public class PluginConstants {
     static final String ON_IDLE_TIMEOUT = "onIdleTimeout";
     static final String ON_TEXT_MESSAGE = "onTextMessage";
     static final String ON_BINARY_MESSAGE = "onBinaryMessage";
+    static final String ON_MESSAGE = "onMessage";
     static final String ON_PING_MESSAGE = "onPing";
     static final String ON_PONG_MESSAGE = "onPong";
     static final String REMOTE_KEY_WORD = "remote";
@@ -84,6 +85,8 @@ public class PluginConstants {
                 "WEBSOCKET_105"),
         INVALID_LISTENER_INIT_PARAMS("`websocket:ListenerConfiguration` not allowed with `http:Listener` "
                 + "as the `websocket:Listener` ", "WEBSOCKET_106"),
+        INVALID_INPUT_FOR_ON_MESSAGE("Invalid parameters `{0}` provided for onMessage remote "
+                + "function", "WEBSOCKET_108"),
         TEMPLATE_CODE_GENERATION_HINT("Template generation for empty service", "WEBSOCKET_107");;
 
         private final String error;