+++ native/src/main/java/io/ballerina/stdlib/websocket/WebSocketUtil.java
@@ -22,9 +22,11 @@ import io.ballerina.runtime.api.Environment;
 import io.ballerina.runtime.api.Future;
 import io.ballerina.runtime.api.Module;
 import io.ballerina.runtime.api.Runtime;
+import io.ballerina.runtime.api.TypeTags;
 import io.ballerina.runtime.api.creators.ErrorCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.types.Type;
+import io.ballerina.runtime.api.types.UnionType;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BDecimal;
@@ -59,9 +61,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
@@ -216,6 +220,22 @@ public class WebSocketUtil {
             return "Unexpected error occurred";
         }
         return err.getMessage();
+    }
+
+    public static BString getBString(byte[] byteArray) {
+        return StringUtils.fromString(
+                new String(byteArray, StandardCharsets.UTF_8));
+    }
+
+    public static boolean hasStringType(Type targetType) {
+        List<Type> memberTypes = ((UnionType) targetType).getMemberTypes();
+        return memberTypes.stream().anyMatch(member -> member.getTag() == TypeTags.STRING_TAG);
+    }
+
+    public static boolean hasByteArrayType(Type targetType) {
+        List<Type> memberTypes = ((UnionType) targetType).getMemberTypes();
+        return memberTypes.stream().anyMatch(member -> member.getTag() == TypeTags.ARRAY_TAG
+                && member.toString().equals(WebSocketConstants.BYTE_ARRAY));
     }
 
     /**